Revert "made required changes"
This reverts commit 28634ea5e54f46d7605f29c34776b106e6bbd9d5.

+++ javacord-api/src/main/java/org/javacord/api/entity/server/Server.java
@@ -2002,7 +2002,7 @@ public interface Server extends DiscordEntity, Nameable, UpdatableFromCache<Serv
      * @param user              The user to ban.
      * @param deleteMessageDays The number of days to delete messages for (0-7).
      * @return A future to check if the ban was successful.
-     * @deprecated Use {@link #banUser(User, int, TimeUnit, String)} instead.
+     * @deprecated Use {@link #banUser(User, long, TimeUnit, String)} instead.
      */
     @Deprecated
     default CompletableFuture<Void> banUser(User user, int deleteMessageDays) {
@@ -2016,7 +2016,7 @@ public interface Server extends DiscordEntity, Nameable, UpdatableFromCache<Serv
      * @param deleteMessageDays The number of days to delete messages for (0-7).
      * @param reason            The reason for the ban.
      * @return A future to check if the ban was successful.
-     * @deprecated Use {@link #banUser(User, int, TimeUnit, String)} instead.
+     * @deprecated Use {@link #banUser(User, long, TimeUnit, String)} instead.
      */
     @Deprecated
     default CompletableFuture<Void> banUser(User user, int deleteMessageDays, String reason) {
@@ -2032,7 +2032,7 @@ public interface Server extends DiscordEntity, Nameable, UpdatableFromCache<Serv
      * @param unit                  The unit of time for the duration.
      * @return A future to check if the ban was successful.
      */
-    default CompletableFuture<Void> banUser(User user, int deleteMessageDuration, TimeUnit unit) {
+    default CompletableFuture<Void> banUser(User user, long deleteMessageDuration, TimeUnit unit) {
         return banUser(user.getId(), deleteMessageDuration, unit);
     }
 
@@ -2046,7 +2046,7 @@ public interface Server extends DiscordEntity, Nameable, UpdatableFromCache<Serv
      * @param reason                The reason for the ban.
      * @return A future to check if the ban was successful.
      */
-    default CompletableFuture<Void> banUser(User user, int deleteMessageDuration, TimeUnit unit, String reason) {
+    default CompletableFuture<Void> banUser(User user, long deleteMessageDuration, TimeUnit unit, String reason) {
         return banUser(user.getId(), deleteMessageDuration, unit, reason);
     }
 
@@ -2099,7 +2099,7 @@ public interface Server extends DiscordEntity, Nameable, UpdatableFromCache<Serv
      * @param userId            The id of the user to ban.
      * @param deleteMessageDays The number of days to delete messages for (0-7).
      * @return A future to check if the ban was successful.
-     * @deprecated Use {@link #banUser(String, int, TimeUnit)} instead.
+     * @deprecated Use {@link #banUser(String, long, TimeUnit)} instead.
      */
     @Deprecated
     default CompletableFuture<Void> banUser(String userId, int deleteMessageDays) {
@@ -2112,7 +2112,7 @@ public interface Server extends DiscordEntity, Nameable, UpdatableFromCache<Serv
      * @param userId            The id of the user to ban.
      * @param deleteMessageDays The number of days to delete messages for (0-7).
      * @return A future to check if the ban was successful.
-     * @deprecated Use {@link #banUser(String, int, TimeUnit)} instead.
+     * @deprecated Use {@link #banUser(String, long, TimeUnit)} instead.
      */
     @Deprecated
     default CompletableFuture<Void> banUser(long userId, int deleteMessageDays) {
@@ -2128,7 +2128,7 @@ public interface Server extends DiscordEntity, Nameable, UpdatableFromCache<Serv
      * @param unit                  The unit of time for the duration.
      * @return A future to check if the ban was successful.
      */
-    default CompletableFuture<Void> banUser(String userId, int deleteMessageDuration, TimeUnit unit) {
+    default CompletableFuture<Void> banUser(String userId, long deleteMessageDuration, TimeUnit unit) {
         return banUser(userId, deleteMessageDuration, unit, null);
     }
 
@@ -2141,7 +2141,7 @@ public interface Server extends DiscordEntity, Nameable, UpdatableFromCache<Serv
      * @param unit                  The unit of time for the duration.
      * @return A future to check if the ban was successful.
      */
-    default CompletableFuture<Void> banUser(long userId, int deleteMessageDuration, TimeUnit unit) {
+    default CompletableFuture<Void> banUser(long userId, long deleteMessageDuration, TimeUnit unit) {
         return banUser(Long.toUnsignedString(userId), deleteMessageDuration, unit);
     }
 
@@ -2174,7 +2174,7 @@ public interface Server extends DiscordEntity, Nameable, UpdatableFromCache<Serv
      * @param deleteMessageDays The number of days to delete messages for (0-7).
      * @param reason            The reason for the ban.
      * @return A future to check if the ban was successful.
-     * @deprecated Use {@link #banUser(String, int, TimeUnit)} instead.
+     * @deprecated Use {@link #banUser(String, long, TimeUnit)} instead.
      */
     @Deprecated
     default CompletableFuture<Void> banUser(String userId, int deleteMessageDays, String reason) {
@@ -2188,7 +2188,7 @@ public interface Server extends DiscordEntity, Nameable, UpdatableFromCache<Serv
      * @param deleteMessageDays The number of days to delete messages for (0-7).
      * @param reason            The reason for the ban.
      * @return A future to check if the ban was successful.
-     * @deprecated Use {@link #banUser(String, int, TimeUnit)} instead.
+     * @deprecated Use {@link #banUser(String, long, TimeUnit)} instead.
      */
     @Deprecated
     default CompletableFuture<Void> banUser(long userId, int deleteMessageDays, String reason) {
@@ -2205,7 +2205,7 @@ public interface Server extends DiscordEntity, Nameable, UpdatableFromCache<Serv
      * @param reason                The reason for the ban.
      * @return A future to check if the ban was successful.
      */
-    CompletableFuture<Void> banUser(String userId, int deleteMessageDuration, TimeUnit unit, String reason);
+    CompletableFuture<Void> banUser(String userId, long deleteMessageDuration, TimeUnit unit, String reason);
 
     /**
      * Bans the given user from the server.
@@ -2217,7 +2217,7 @@ public interface Server extends DiscordEntity, Nameable, UpdatableFromCache<Serv
      * @param reason                The reason for the ban.
      * @return A future to check if the ban was successful.
      */
-    default CompletableFuture<Void> banUser(long userId, int deleteMessageDuration, TimeUnit unit, String reason) {
+    default CompletableFuture<Void> banUser(long userId, long deleteMessageDuration, TimeUnit unit, String reason) {
         return banUser(Long.toUnsignedString(userId), deleteMessageDuration, unit, reason);
     }
 
@@ -2230,7 +2230,7 @@ public interface Server extends DiscordEntity, Nameable, UpdatableFromCache<Serv
      * @return A future to check if the ban was successful.
      */
     default CompletableFuture<Void> banUser(String userId, Duration duration, String reason) {
-        return banUser(userId, (int) duration.getSeconds(), TimeUnit.SECONDS, reason);
+        return banUser(userId, duration.getSeconds(), TimeUnit.SECONDS, reason);
     }
 
     /**
